backend/requirements.txt
annotated-types==0.7.0
bcrypt==4.0.1
blinker==1.9.0
cffi==2.0.0
click==8.1.8
cryptography==46.0.5
dnspython==2.7.0
email-validator==2.3.0
flask==3.1.2
Flask-Bcrypt==1.0.1
flask-cors==6.0.2
Flask-JWT-Extended==4.7.1
greenlet==3.2.4
idna==3.11
importlib-metadata==8.7.1
itsdangerous==2.2.0
jinja2==3.1.6
markupsafe==3.0.3
passlib==1.7.4
pycparser==2.23
pydantic==2.12.5
pydantic-core==2.41.5
pyjwt==2.11.0
pymysql==1.1.2
python-dotenv==1.2.1
sqlalchemy==2.0.46
typing-extensions==4.15.0
typing-inspection==0.4.2
werkzeug==3.1.5
zipp==3.23.0
